requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
pymupdf
pybind11
requests
numpy
google-generativeai
python-multipart